refactor: move ValueChangeMode logic to InternalFieldBase
